+++ lambda/chalicelib/routes/responseEdit.py
@@ -8,7 +8,10 @@ from chalicelib.models import (
 )
 from bson.objectid import ObjectId
 from pydash.objects import get, set_
-from chalicelib.routes.responseIpnListener import mark_successful_payment, send_email_receipt
+from chalicelib.routes.responseIpnListener import (
+    mark_successful_payment,
+    send_email_receipt,
+)
 import dateutil
 
 
@@ -105,9 +108,11 @@ def response_edit_admin_info(responseId):
 def response_checkin(formId, responseId):
     pass
 
+
 def _parse_email_parameters(json_body):
     email_template_id = json_body.get("emailTemplateId", None)
     return email_template_id
+
 
 """
 Adds a payment. (POST)
@@ -126,6 +131,8 @@ JSON body parameters:
     "emailTemplateId": "id" # id corresponding to the template in formOptions.confirmationEmailTemplates to use when sending the confirmation email.
 }
 """
+
+
 def response_add_payment(responseId):
     from ..main import app
 
@@ -157,12 +164,13 @@ def response_add_payment(responseId):
         date=date,
         send_email=send_email,
         notes=notes,
-        email_template_id=email_template_id
+        email_template_id=email_template_id,
     )
     response.save()
     return {
         "res": {"success": True, "paid": paid, "response": serialize_model(response)}
     }
+
 
 """
 Sends an email.
@@ -173,16 +181,17 @@ JSON body parameters:
     "emailTemplateId": "id" # id corresponding to the template in formOptions.confirmationEmailTemplates to use when sending the confirmation email.
 }
 """
+
+
 def response_send_email(responseId):
     from ..main import app
+
     response = Response.objects.get({"_id": ObjectId(responseId)})
     app.check_permissions(response.form, ["Responses_Edit", "Responses_SendEmail"])
     email_template_id = _parse_email_parameters(app.current_request.json_body)
     send_email_receipt(response, response.form.formOptions, email_template_id)
     response.save()
-    return {
-        "res": {"success": True, "response": serialize_model(response)}
-    }
+    return {"res": {"success": True, "response": serialize_model(response)}}
 
 
 def response_delete(responseId):