+++ lambda/tests/integration/test_formSubmit.py
@@ -16,6 +16,7 @@ from chalicelib.models import Form, Response, CCAvenueConfig, serialize_model
 from bson.objectid import ObjectId
 import time
 import mock
+import uuid
 
 ONE_SUBMITRES = {
     "paid": False,
@@ -35,6 +36,7 @@ ONE_SUBMITRES = {
         ],
         "total": 0.5,
     },
+    "amount_owed_cents": 50,
     "paymentMethods": {
         "paypal_classic": {
             "address1": "123",
@@ -119,7 +121,6 @@ class FormSubmit(BaseTestCase):
             ],
             "total": 100.0,
         }
-        print(submit_res["paymentInfo"])
         self.assertEqual(submit_res["paymentInfo"], expected_paymentInfo)
         response = Response.objects.get({"_id": ObjectId(responseId)})
         self.assertEqual(response.paymentInfo, expected_paymentInfo)
@@ -266,6 +267,7 @@ class FormSubmit(BaseTestCase):
         self.assertEqual(response["paid"], True)
         self.assertEqual(response["amount_paid"], "0.5")
         self.assertEqual(response["paymentInfo"]["total"], 0.5)
+        self.assertEqual(response["amount_owed_cents"], 50)
 
         new_data = dict(ONE_FORMDATA, children=[{}])
 
@@ -282,6 +284,7 @@ class FormSubmit(BaseTestCase):
         self.assertEqual(response["paid"], False)
         self.assertEqual(response["amount_paid"], "0.5")
         self.assertEqual(response["paymentInfo"]["total"], 25.5)
+        self.assertEqual(response["amount_owed_cents"], 2550)
 
         # Pay response.
         response = Response.objects.get({"_id": ObjectId(responseId)})
@@ -771,6 +774,7 @@ class FormSubmit(BaseTestCase):
                     ],
                     "total": 0.5,
                 },
+                "amount_owed_cents": 50,
                 "paymentMethods": {
                     "paypal_classic": {
                         "address1": "123",
@@ -898,6 +902,7 @@ class FormSubmit(BaseTestCase):
                     ],
                     "total": 0.5,
                 },
+                "amount_owed_cents": 50,
                 "paymentMethods": {
                     "paypal_classic": {
                         "address1": "123",
@@ -1009,8 +1014,9 @@ class FormSubmit(BaseTestCase):
         )
         self.assertEqual(response["paid"], True)
         self.assertEqual(response["amount_paid"], "0.5")
+        self.assertEqual(response["amount_paid_cents"], 50)
         self.assertEqual(len(response["email_trail"]), 1)
-    
+
     @mock.patch("boto3.client")
     def test_mark_successful_payment_custom_email_template(self, mock_boto_client):
         formOptions = {
@@ -1020,10 +1026,8 @@ class FormSubmit(BaseTestCase):
                     "confirmationEmailInfo": {
                         "toField": "email",
                         "subject": "subject 123",
-                        "template": {
-                            "html": "test123"
-                        }
-                    }
+                        "template": {"html": "test123"},
+                    },
                 }
             ]
         }
@@ -1042,7 +1046,7 @@ class FormSubmit(BaseTestCase):
             amount=0.5,
             currency="USD",
             id="payment1",
-            email_template_id="template1"
+            email_template_id="template1",
         )
         response.save()
         self.assertEqual(paid, True)
@@ -1158,9 +1162,9 @@ class FormSubmit(BaseTestCase):
     def test_submit_form_ccavenue(self):
         self.formId = self.create_form()
         access_code = "ACCESSKODE"
-        merchant_id = "123456"
+        merchant_id = str(uuid.uuid4())
         SECRET_working_key = "abcdefgh"
-        CCAvenueConfig(
+        config = CCAvenueConfig(
             access_code=access_code,
             merchant_id=merchant_id,
             SECRET_working_key=SECRET_working_key,