+++ scripts/src/admin/ResponseTable/ResponseTableView.tsx
@@ -1,5 +1,6 @@
-import React from "react";
+import React, { Component } from "react";
 import ReactTable from "react-table";
+import classnames from "classnames";
 import { IResponse } from "../../store/responses/types";
 import { IDataOptionView, IRenderedForm } from "../FormEdit/FormEdit.d";
 import downloadCSV from "./downloadCSV";
@@ -7,6 +8,28 @@ import { filterCaseInsensitive } from "./filters";
 import Modal from "react-responsive-modal";
 import ResponseDetail from "./ResponseDetail";
 import { createHeadersAndDataFromDataOption } from "../util/dataOptionUtil";
+import { NavLink, Link } from "react-router-dom";
+import {
+  ContextMenu,
+  MenuItem,
+  ContextMenuTrigger,
+  SubMenu
+} from "react-contextmenu";
+
+const ResponseContextMenu = ({ id, viewLink, editLink, onInspect }) => (
+  <ContextMenu id={id}>
+    <Link to={viewLink}>
+      <MenuItem>View</MenuItem>
+    </Link>
+
+    <Link to={editLink}>
+      <MenuItem>Edit</MenuItem>
+    </Link>
+
+    <MenuItem divider />
+    <MenuItem onClick={e => onInspect()}>Inspect</MenuItem>
+  </ContextMenu>
+);
 
 interface IResponseTableViewProps {
   responses: IResponse[];
@@ -49,6 +72,18 @@ export default (props: IResponseTableViewProps) => {
         defaultSorted={[{ id: "DATE_LAST_MODIFIED", desc: true }]}
         defaultFiltered={[{ id: "PAID", value: "all" }]}
         defaultFilterMethod={filterCaseInsensitive}
+        TdComponent={({ toggleSort, className, children, ...rest }) => {
+          const id = children.props.row._original.ID;
+          return (
+            <div
+              className={classnames("rt-td", className)}
+              role="gridcell"
+              {...rest}
+            >
+              <ContextMenuTrigger id={id}>{children}</ContextMenuTrigger>
+            </div>
+          );
+        }}
         getTdProps={(state, rowInfo, column, instance) => {
           return {
             onClick: e => {
@@ -72,6 +107,18 @@ export default (props: IResponseTableViewProps) => {
           </div>
         </div>
       </Modal>
+      {props.responses.map(response => {
+        const id = response._id.$oid;
+        return (
+          <ResponseContextMenu
+            key={id}
+            id={id}
+            viewLink={`/v2/forms/${props.renderedForm._id.$oid}?responseId=${id}&mode=view`}
+            editLink={`/v2/forms/${props.renderedForm._id.$oid}?responseId=${id}&mode=edit`}
+            onInspect={() => props.displayResponseDetail(id)}
+          />
+        );
+      })}
     </div>
   );
 };