+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 3.28.0
+- Add cff_createdBetween to restrict payment items by deadline
+- Allow specifiedShowFields to override uiSchema with keys with a CFF_uiSchema prefix
+
+## 3.27.0
+- Add right-click context menu for responses view (#199)
+- Add form summary page (#208)
+  - This also requires a db change that adds the amount_paid_cents and amount_owed_cents fields (lambda/tools/addPaymentFields.py)
+
 ## 3.26.1
 - Fix: Add React Table css to all pages